+++ src/app/builder/create/page.tsx
@@ -10,8 +10,8 @@ import ActionButton from '@/features/build/components/ActionButton'
 import { initialBuildState } from '@/features/build/lib'
 import useBuildActions from '@/features/build/hooks/useBuildActions'
 import BuilderPage from '@/features/build/components/BuilderPage'
-import ArmorCalculatorDialog from '@/features/build/components/ArmorCalculatorDialog'
 import { BuildState } from '@/features/build/types'
+import { BuildSuggestionsDialog } from '@/features/build/components/BuildSuggestionsDialog'
 
 export default function Page() {
   const isClient = useIsClient()
@@ -31,11 +31,11 @@ export default function Page() {
   const buildContainerRef = useRef<HTMLDivElement>(null)
   const detailedViewContainerRef = useRef<HTMLDivElement>(null)
 
-  const [showArmorCalculator, setShowArmorCalculator] = useState(false)
+  const [showBuildSuggestions, setShowBuildSuggestions] = useState(false)
 
-  function handleSelectArmorSuggestion(newBuildState: BuildState) {
+  function handleApplySuggestions(newBuildState: BuildState) {
     setNewBuildState(newBuildState)
-    setShowArmorCalculator(false)
+    setShowBuildSuggestions(false)
   }
 
   if (!isClient) return null
@@ -49,11 +49,11 @@ export default function Page() {
         subtitle="Create your builds and share them with your friends and the community."
       />
 
-      <ArmorCalculatorDialog
+      <BuildSuggestionsDialog
         buildState={dbBuildState}
-        open={showArmorCalculator}
-        onClose={() => setShowArmorCalculator(false)}
-        onSelectArmorSuggestion={handleSelectArmorSuggestion}
+        open={showBuildSuggestions}
+        onClose={() => setShowBuildSuggestions(false)}
+        onApplySuggestions={handleApplySuggestions}
       />
 
       <BuilderPage
@@ -70,8 +70,8 @@ export default function Page() {
           <>
             <SaveBuildButton buildState={dbBuildState} editMode={false} />
 
-            <ActionButton.ArmorCalculator
-              onClick={() => setShowArmorCalculator(true)}
+            <ActionButton.BuildSuggestions
+              onClick={() => setShowBuildSuggestions(true)}
             />
 
             <ActionButton.ShowDetailedView